walls
top and bottom walls now cost money. If the wall button is clicked again
the wall is replaced by an upgraded wall and the old one deletes itself.
must add code to reset the wall variables in gamemanager if a wall is
destroyed so the player can build new ones.